throughput-sync: check trace existence before reading it

+++ throughput-sync/throughput-sync.sh
@@ -185,5 +185,7 @@ print_save_agg_thr $file_name
 cd ..
 
 # rm work dir
-cp -f results-${sched}/trace .
+if [ -f results-${sched}/trace ]; then
+    cp -f results-${sched}/trace .
+fi
 rm -rf results-${sched}
